Fix date issues with the description updater (#4)
* Hack fix to increase date precision

Since discord.py does not have a task scheduler based on a date, we makeshift our own. It will loop every second (which only actually starts that timer once the function exits. The loop will not run concurrently) and if the date equals the currently know date state then await the difference in seconds until tomorrow, then exit. When the date does not match it will instead change the channel topic and then set the state date to the current day, and begin again.

* Prevent KeyError by out of range days

When the event starts and ends it will create a KeyError from a non-existent day in day_themes (i.e. now_day - 1 == 0 OR now_day + 1 == 32). When the event passes 2 days after the end (November 2nd) there will be no valid day themes, so change the topic to reflect that.

* Only count day if in ink_month

* lang update

* Properly track months

* Syntax errors

* Transition over to library time arg